Onboarding layout added, still missing assets for background to show
