--- app/src/main/java/com/shiokority/shiokoritypay/view/MainActivity.kt
@@ -1,40 +1,0 @@
-package com.shiokority.shiokoritypay.view
-
-import android.os.Bundle
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
-import com.shiokority.shiokoritypay.R
-import com.shiokority.shiokoritypay.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-     binding = ActivityMainBinding.inflate(layoutInflater)
-     setContentView(binding.root)
-
-        val navView: BottomNavigationView = binding.navView
-
-        val navController = findNavController(R.id.nav_host_fragment_activity_main)
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
-        val appBarConfiguration = AppBarConfiguration(setOf(
-//            R.id.navigation_home, R.id.navigation_history, R.id.navigation_notifications))
-            R.id.navigation_home,
-            R.id.navigation_payment,
-            R.id.navigation_pay,
-            R.id.navigation_history,
-            R.id.navigation_pay,
-            R.id.navigation_profile
-        ))
-        setupActionBarWithNavController(navController, appBarConfiguration)
-        navView.setupWithNavController(navController)
-    }
-}
