+++ app/src/main/java/com/shiokority/shiokoritypay/view/authentication/login/LoginViewModel.kt
@@ -1,0 +1,4 @@
+package com.shiokority.shiokoritypay.view.authentication.login
+
+class LoginViewModel {
+}
