+++ app/src/main/java/com/shiokority/shiokoritypay/view/authentication/register/RegisterAcitivity.kt
@@ -1,0 +1,4 @@
+package com.shiokority.shiokoritypay.view.authentication.register
+
+class RegisterAcitivity {
+}
