+++ app/src/main/java/com/shiokority/shiokoritypay/view/history/MyPaymentHistoryRecyclerViewAdapter.java
@@ -1,0 +1,61 @@
+package com.shiokority.shiokoritypay.view.history;
+
+import androidx.recyclerview.widget.RecyclerView;
+
+import android.view.LayoutInflater;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import com.shiokority.shiokoritypay.view.history.placeholder.PlaceholderContent.PlaceholderItem;
+import com.shiokority.shiokoritypay.databinding.FragmentItemBinding;
+
+import java.util.List;
+
+/**
+ * {@link RecyclerView.Adapter} that can display a {@link PlaceholderItem}.
+ * TODO: Replace the implementation with code for your data type.
+ */
+public class MyPaymentHistoryRecyclerViewAdapter extends RecyclerView.Adapter<MyPaymentHistoryRecyclerViewAdapter.ViewHolder> {
+
+    private final List<PlaceholderItem> mValues;
+
+    public MyPaymentHistoryRecyclerViewAdapter(List<PlaceholderItem> items) {
+        mValues = items;
+    }
+
+    @Override
+    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+
+        return new ViewHolder(FragmentItemBinding.inflate(LayoutInflater.from(parent.getContext()), parent, false));
+
+    }
+
+    @Override
+    public void onBindViewHolder(final ViewHolder holder, int position) {
+        holder.mItem = mValues.get(position);
+        holder.mIdView.setText(mValues.get(position).id);
+        holder.mContentView.setText(mValues.get(position).content);
+    }
+
+    @Override
+    public int getItemCount() {
+        return mValues.size();
+    }
+
+    public class ViewHolder extends RecyclerView.ViewHolder {
+        public final TextView mIdView;
+        public final TextView mContentView;
+        public PlaceholderItem mItem;
+
+        public ViewHolder(FragmentItemBinding binding) {
+            super(binding.getRoot());
+            mIdView = binding.itemNumber;
+            mContentView = binding.content;
+        }
+
+        @Override
+        public String toString() {
+            return super.toString() + " '" + mContentView.getText() + "'";
+        }
+    }
+}
