+++ app/src/main/java/com/shiokority/shiokoritypay/view/onboarding/OnboardingAdapter.kt
@@ -1,0 +1,44 @@
+package com.shiokority.shiokoritypay.view.onboarding
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.annotation.DrawableRes
+import androidx.recyclerview.widget.RecyclerView
+import com.shiokority.shiokoritypay.R
+
+class OnboardingAdapter(private val onboardingItems: List<OnboardingItem>) :
+    RecyclerView.Adapter<OnboardingAdapter.OnboardingViewHolder>() {
+
+    inner class OnboardingViewHolder(view: View) : RecyclerView.ViewHolder(view) {
+        private val imageOnboarding: ImageView = view.findViewById(R.id.imageOnboarding)
+        private val textTitle: TextView = view.findViewById(R.id.textTitle)
+        private val textDescription: TextView = view.findViewById(R.id.textDescription)
+
+        fun bind(onboardingItem: OnboardingItem) {
+            imageOnboarding.setImageResource(onboardingItem.image)
+            textTitle.text = onboardingItem.title
+            textDescription.text = onboardingItem.description
+        }
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): OnboardingViewHolder {
+        return OnboardingViewHolder(
+            LayoutInflater.from(parent.context).inflate(R.layout.item_onboarding_page, parent, false)
+        )
+    }
+
+    override fun onBindViewHolder(holder: OnboardingViewHolder, position: Int) {
+        holder.bind(onboardingItems[position])
+    }
+
+    override fun getItemCount(): Int = onboardingItems.size
+}
+
+data class OnboardingItem(
+    @DrawableRes val image: Int,
+    val title: String,
+    val description: String
+)
