+++ app/src/main/java/com/shiokority/shiokoritypay/view/pay/PayViewModel.kt
@@ -1,0 +1,13 @@
+package com.shiokority.shiokoritypay.view.pay
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class PayViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is Pay Fragment"
+    }
+    val text: LiveData<String> = _text
+}
