+++ app/src/main/java/com/shiokority/shiokoritypay/view/profile/ProfileViewModel.kt
@@ -1,0 +1,13 @@
+package com.shiokority.shiokoritypay.view.profile
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class ProfileViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is Profile Fragment"
+    }
+    val text: LiveData<String> = _text
+}
