--- app/src/main/java/com/shiokority/shiokoritypay/view/ui/history/HistoryFragment.kt
@@ -1,41 +1,0 @@
-package com.shiokority.shiokoritypay.view.ui.history
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import com.shiokority.shiokoritypay.databinding.FragmentHomeBinding
-
-class HistoryFragment : Fragment() {
-
-private var _binding: FragmentHomeBinding? = null
-  // This property is only valid between onCreateView and
-  // onDestroyView.
-  private val binding get() = _binding!!
-
-  override fun onCreateView(
-    inflater: LayoutInflater,
-    container: ViewGroup?,
-    savedInstanceState: Bundle?
-  ): View {
-    val historyViewModel =
-            ViewModelProvider(this).get(HistoryViewModel::class.java)
-
-    _binding = FragmentHomeBinding.inflate(inflater, container, false)
-    val root: View = binding.root
-
-    val textView: TextView = binding.textHome
-      historyViewModel.text.observe(viewLifecycleOwner) {
-      textView.text = it
-    }
-    return root
-  }
-
-override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
