+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ui.dashboard.DashboardFragment" >
+    tools:context=".view.dashboard.DashboardFragment" >
 
     <TextView
         android:id="@+id/nav_dashboard"
