+++ app/src/main/res/layout/fragment_notifications.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ui.notifications.NotificationsFragment" >
+    tools:context=".view.notifications.NotificationsFragment" >
 
     <TextView
         android:id="@+id/text_notifications"
