+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,6 @@
     <string name="text_login_to_your_account">Login to your Account</string>
     <string name="text_forgot_password">Forgot Password?</string>
     <string name="toggle_password_visibility">Show Password</string>
+    <string name="button_next">Next</string>
+    <string name="tab_layout_description">describe onboarding content</string>
 </resources>
